+++ Blazicons.Generating/AssemblyInfo.cs
@@ -1,1 +1,4 @@
-﻿[assembly: CLSCompliant(false)] // Microsoft.CodeAnalysis
+﻿using System.Runtime.CompilerServices;
+
+[assembly: CLSCompliant(false)] // Microsoft.CodeAnalysis
+[assembly: InternalsVisibleTo("Blazicons.Generating.UnitTests")]