+++ Blazicons.Generating/Internals/SvgColorType.cs
@@ -1,0 +1,11 @@
+﻿namespace Blazicons.Generating.Internals;
+
+internal enum SvgColorType
+{
+    None,
+    SingleFill,
+    SingleStroke,
+    SingleColor,
+    SingleShared,
+    Multiple,
+}