+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 These are the list of changes to each version that has been released.
 
+## 2.0
+- [x] Added support for generating code compatible with Blazicons v2.0
+- [x] Enhanced color parsing functionality
+
 ## 1.1
 - [x] Fix for cross-platform compatibility